IOS-1943 Extracted UserWalletCell View Model
